Split audio provider helpers
